INSA_Cybersecurity_Group_13-main/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.1
python-nmap==0.7.1
requests==2.31.0
 Flask
 Flask-Cors
 Flask-SocketIO
 psutil
 Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
psutil==5.9.5
xmltodict==0.13.0